Move bottom navbar scaffold into components dir
