+++ app/src/main/java/com/simply/birthdayapp/presentation/ui/components/BottomNavBarScaffold.kt
@@ -1,4 +1,4 @@
-package com.simply.birthdayapp.presentation.ui.screens.main
+package com.simply.birthdayapp.presentation.ui.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
