+++ app/src/main/java/com/simply/birthdayapp/presentation/ui/screens/main/MainScreen.kt
@@ -3,6 +3,7 @@ package com.simply.birthdayapp.presentation.ui.screens.main
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.rememberNavController
 import com.simply.birthdayapp.presentation.navigation.BottomNavigation
+import com.simply.birthdayapp.presentation.ui.components.BottomNavBarScaffold
 
 @Composable
 fun MainScreen() {
